Exe can now use resource pack

+++ trunk/KeepItEmpty/Level.cpp
@@ -80,14 +80,14 @@ void Level::Update()
 							if(m_ppElements[cpt]->GetState() == Dead && !m_ppElements[i]->GetAlreadyMultiply())
 							{
 								m_ppElements[i]->SetAlreadyMultiply(true);
-								m_ppElements[cpt] = new Element(m_pGame, m_ppElements[i]->GetPosX(), m_ppElements[i]->GetPosY(),m_pGame->nRadius, m_pGame->fSpeed);
+								m_ppElements[cpt] = new Element(m_pGame, (int)m_ppElements[i]->GetPosX(), (int)m_ppElements[i]->GetPosY(),m_pGame->nRadius, m_pGame->fSpeed);
 								for(int cpt2 = cpt+1;cpt2 < cpt+12;cpt2++)
 								{
 									if(cpt2 < m_nElements)
 									{
 										if(m_ppElements[cpt2]->GetState() == Dead)
 										{
-											m_ppElements[cpt2] = new Element(m_pGame, m_ppElements[i]->GetPosX(), m_ppElements[i]->GetPosY(),m_pGame->nRadius, m_pGame->fSpeed);
+											m_ppElements[cpt2] = new Element(m_pGame, (int)m_ppElements[i]->GetPosX(), (int)m_ppElements[i]->GetPosY(),m_pGame->nRadius, m_pGame->fSpeed);
 										}
 									}
 								}
